Break up caching a bit
