+++ .github/workflows/pytest.yml
@@ -19,7 +19,6 @@ jobs:
           python-version: "3.6"
       - name: Install dependencies
         run: |
-          sudo apt-get install libsnappy-dev
           python -m pip install --upgrade pip
           pip install -e .[test]
       - name: Run tests
